requirements.txt
mistralai
python-dotenv
openai
PyQt6
pyaudio
numpy
pygame
websockets